requirements.txt
openai
numpy
sounddevice
pydub
elevenlabs
pynput
pyserial
